Rename filter as lens to avoid overwriting inbuilt filter function
